requirements.txt
astroid==3.3.5
certifi==2024.8.30
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dill==0.3.9
dnspython==2.6.1
Flask==2.1.2
fonttools==4.57.0
greenlet==3.2.0
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.1
mccabe==0.7.0
neo4j==5.25.0
numpy==2.2.4
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.6
pylint==3.3.1
pymongo==4.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.27.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
SQLAlchemy==2.0.40
threadpoolctl==3.6.0
tomlkit==0.13.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
Werkzeug==2.0.3
